apart headline listing into component.
